requirements.txt
openai>=0.27.0
tiktoken>=0.3.0
PyYAML==6.0
pymongo==4.3.3
pyrogram
TgCrypto